[fix] fix bug when partition_id exceeds integer range in spark load (#9073)

+++ fe/spark-dpp/src/main/java/org/apache/doris/load/loadv2/dpp/SparkDpp.java
@@ -248,7 +248,7 @@ public final class SparkDpp implements java.io.Serializable {
                             LOG.warn("invalid bucket key:" + curBucketKey);
                             continue;
                         }
-                        int partitionId = Integer.parseInt(bucketKey[0]);
+                        long partitionId = Long.parseLong(bucketKey[0]);
                         int bucketId = Integer.parseInt(bucketKey[1]);
                         dstPath = String.format(pathPattern, tableId, partitionId, indexMeta.indexId,
                                 bucketId, indexMeta.schemaHash);
